chore: forge fmt

+++ contracts/v0.2/test/FeedRouter.t.sol
@@ -16,109 +16,109 @@ contract FeedRouterTest is Test {
     }
 
     function test_AddProxy() public {
-	string[] memory feedNames_ = new string[](2);
-	address[] memory proxyAddresses_ = new address[](2);
+        string[] memory feedNames_ = new string[](2);
+        address[] memory proxyAddresses_ = new address[](2);
 
-	address btcUsdt = makeAddr("btc-usdt");
-	address ethUsdt = makeAddr("eth-usdt");
+        address btcUsdt = makeAddr("btc-usdt");
+        address ethUsdt = makeAddr("eth-usdt");
 
-	feedNames_[0] = "btc-usdt";
-	feedNames_[1] = "eth-usdt";
+        feedNames_[0] = "btc-usdt";
+        feedNames_[1] = "eth-usdt";
 
-	proxyAddresses_[0] = btcUsdt;
-	proxyAddresses_[1] = ethUsdt;
+        proxyAddresses_[0] = btcUsdt;
+        proxyAddresses_[1] = ethUsdt;
 
-	vm.expectEmit(true, true, true, true);
-	emit ProxyAdded(feedNames_[0], proxyAddresses_[0]);
-	vm.expectEmit(true, true, true, true);
-	emit ProxyAdded(feedNames_[1], proxyAddresses_[1]);
-	feedRouter.updateProxyBulk(feedNames_, proxyAddresses_);
+        vm.expectEmit(true, true, true, true);
+        emit ProxyAdded(feedNames_[0], proxyAddresses_[0]);
+        vm.expectEmit(true, true, true, true);
+        emit ProxyAdded(feedNames_[1], proxyAddresses_[1]);
+        feedRouter.updateProxyBulk(feedNames_, proxyAddresses_);
 
-	assertEq(feedRouter.feedToProxies(feedNames_[0]), btcUsdt);
-	assertEq(feedRouter.feedToProxies(feedNames_[1]), ethUsdt);
-	assertEq(true, compareArrays(feedNames_, feedRouter.getFeedNames()));
+        assertEq(feedRouter.feedToProxies(feedNames_[0]), btcUsdt);
+        assertEq(feedRouter.feedToProxies(feedNames_[1]), ethUsdt);
+        assertEq(true, compareArrays(feedNames_, feedRouter.getFeedNames()));
     }
 
     function test_UpdateProxy() public {
-	string[] memory feedNames_ = new string[](1);
-	address[] memory proxyAddressesOld_ = new address[](1);
-	address[] memory proxyAddressesNew_ = new address[](1);
+        string[] memory feedNames_ = new string[](1);
+        address[] memory proxyAddressesOld_ = new address[](1);
+        address[] memory proxyAddressesNew_ = new address[](1);
 
-	address btcUsdtOld = makeAddr("btc-usdt-old");
-	address btcUsdtNew = makeAddr("btc-usdt-new");
-	feedNames_[0] = "btc-usdt";
+        address btcUsdtOld = makeAddr("btc-usdt-old");
+        address btcUsdtNew = makeAddr("btc-usdt-new");
+        feedNames_[0] = "btc-usdt";
 
-	proxyAddressesOld_[0] = btcUsdtOld;
-	proxyAddressesNew_[0] = btcUsdtNew;
+        proxyAddressesOld_[0] = btcUsdtOld;
+        proxyAddressesNew_[0] = btcUsdtNew;
 
-	vm.expectEmit(true, true, true, true);
-	emit ProxyAdded(feedNames_[0], proxyAddressesOld_[0]);
-	feedRouter.updateProxyBulk(feedNames_, proxyAddressesOld_);
-	assertEq(btcUsdtOld, feedRouter.feedToProxies(feedNames_[0]));
-	assertEq(feedRouter.feedToProxies(feedNames_[0]), btcUsdtOld);
+        vm.expectEmit(true, true, true, true);
+        emit ProxyAdded(feedNames_[0], proxyAddressesOld_[0]);
+        feedRouter.updateProxyBulk(feedNames_, proxyAddressesOld_);
+        assertEq(btcUsdtOld, feedRouter.feedToProxies(feedNames_[0]));
+        assertEq(feedRouter.feedToProxies(feedNames_[0]), btcUsdtOld);
 
-	vm.expectEmit(true, true, true, true);
-	emit ProxyUpdated(feedNames_[0], proxyAddressesNew_[0]);
-	feedRouter.updateProxyBulk(feedNames_, proxyAddressesNew_);
-	assertEq(btcUsdtNew, feedRouter.feedToProxies(feedNames_[0]));
-	assertEq(feedRouter.feedToProxies(feedNames_[0]), btcUsdtNew);
+        vm.expectEmit(true, true, true, true);
+        emit ProxyUpdated(feedNames_[0], proxyAddressesNew_[0]);
+        feedRouter.updateProxyBulk(feedNames_, proxyAddressesNew_);
+        assertEq(btcUsdtNew, feedRouter.feedToProxies(feedNames_[0]));
+        assertEq(feedRouter.feedToProxies(feedNames_[0]), btcUsdtNew);
     }
 
     function test_UpdateNotEqualLength() public {
-	string[] memory feedNames_ = new string[](2);
-	address[] memory proxyAddresses_ = new address[](1);
+        string[] memory feedNames_ = new string[](2);
+        address[] memory proxyAddresses_ = new address[](1);
 
-	// feedNames_ is longer than proxyAddresses_ -> FAIL
-	vm.expectRevert("invalid input");
-	feedRouter.updateProxyBulk(feedNames_, proxyAddresses_);
+        // feedNames_ is longer than proxyAddresses_ -> FAIL
+        vm.expectRevert("invalid input");
+        feedRouter.updateProxyBulk(feedNames_, proxyAddresses_);
     }
 
     function test_UpdateInvalidProxy() public {
-	string[] memory feedNames_ = new string[](1);
-	address[] memory proxyAddresses_ = new address[](1);
+        string[] memory feedNames_ = new string[](1);
+        address[] memory proxyAddresses_ = new address[](1);
 
-	feedNames_[0] = "btc-usdt";
-	proxyAddresses_[0] = address(0);
+        feedNames_[0] = "btc-usdt";
+        proxyAddresses_[0] = address(0);
 
-	// proxy address is 0 -> FAIL
-	vm.expectRevert(FeedRouter.InvalidProxyAddress.selector);
-	feedRouter.updateProxyBulk(feedNames_, proxyAddresses_);
+        // proxy address is 0 -> FAIL
+        vm.expectRevert(FeedRouter.InvalidProxyAddress.selector);
+        feedRouter.updateProxyBulk(feedNames_, proxyAddresses_);
     }
 
     function test_RemoveProxy() public {
-	string[] memory feedNamesAdd_ = new string[](2);
-	address[] memory proxyAddresses_ = new address[](2);
+        string[] memory feedNamesAdd_ = new string[](2);
+        address[] memory proxyAddresses_ = new address[](2);
 
-	address btcUsdt = makeAddr("btc-usdt");
-	address ethUsdt = makeAddr("eth-usdt");
+        address btcUsdt = makeAddr("btc-usdt");
+        address ethUsdt = makeAddr("eth-usdt");
 
-	feedNamesAdd_[0] = "btc-usdt";
-	feedNamesAdd_[1] = "eth-usdt";
+        feedNamesAdd_[0] = "btc-usdt";
+        feedNamesAdd_[1] = "eth-usdt";
 
-	proxyAddresses_[0] = btcUsdt;
-	proxyAddresses_[1] = ethUsdt;
+        proxyAddresses_[0] = btcUsdt;
+        proxyAddresses_[1] = ethUsdt;
 
-	// add proxies
-	vm.expectEmit(true, true, true, true);
-	emit ProxyAdded(feedNamesAdd_[0], proxyAddresses_[0]);
-	feedRouter.updateProxyBulk(feedNamesAdd_, proxyAddresses_);
-	assertEq(feedRouter.feedToProxies(feedNamesAdd_[0]), btcUsdt);
-	assertEq(feedRouter.feedToProxies(feedNamesAdd_[1]), ethUsdt);
+        // add proxies
+        vm.expectEmit(true, true, true, true);
+        emit ProxyAdded(feedNamesAdd_[0], proxyAddresses_[0]);
+        feedRouter.updateProxyBulk(feedNamesAdd_, proxyAddresses_);
+        assertEq(feedRouter.feedToProxies(feedNamesAdd_[0]), btcUsdt);
+        assertEq(feedRouter.feedToProxies(feedNamesAdd_[1]), ethUsdt);
 
-	// remove btc-usdt proxy
-	string[] memory feedNamesRemove_ = new string[](1);
-	feedNamesRemove_[0] = feedNamesAdd_[0];
-	vm.expectEmit(true, true, true, true);
-	emit ProxyRemoved(feedNamesRemove_[0], proxyAddresses_[0]);
-	feedRouter.removeProxyBulk(feedNamesRemove_);
+        // remove btc-usdt proxy
+        string[] memory feedNamesRemove_ = new string[](1);
+        feedNamesRemove_[0] = feedNamesAdd_[0];
+        vm.expectEmit(true, true, true, true);
+        emit ProxyRemoved(feedNamesRemove_[0], proxyAddresses_[0]);
+        feedRouter.removeProxyBulk(feedNamesRemove_);
 
-	// expect that only eth-usdt is left
-	string[] memory expectedFeedNames_ = new string[](1);
-	expectedFeedNames_[0] = feedNamesAdd_[1];
-	assertEq(true, compareArrays(expectedFeedNames_, feedRouter.getFeedNames()));
+        // expect that only eth-usdt is left
+        string[] memory expectedFeedNames_ = new string[](1);
+        expectedFeedNames_[0] = feedNamesAdd_[1];
+        assertEq(true, compareArrays(expectedFeedNames_, feedRouter.getFeedNames()));
 
-	assertEq(feedRouter.feedToProxies(feedNamesAdd_[0]), address(0));
-	assertEq(feedRouter.feedToProxies(feedNamesAdd_[1]), ethUsdt);
+        assertEq(feedRouter.feedToProxies(feedNamesAdd_[0]), address(0));
+        assertEq(feedRouter.feedToProxies(feedNamesAdd_[1]), ethUsdt);
     }
 
     function compareArrays(string[] memory array1, string[] memory array2) private pure returns (bool) {
@@ -126,9 +126,9 @@ contract FeedRouterTest is Test {
             return false;
         }
 
-        for (uint i = 0; i < array1.length; i++) {
+        for (uint256 i = 0; i < array1.length; i++) {
             bool found = false;
-            for (uint j = 0; j < array2.length; j++) {
+            for (uint256 j = 0; j < array2.length; j++) {
                 if (keccak256(abi.encodePacked(array1[i])) == keccak256(abi.encodePacked(array2[j]))) {
                     found = true;
                     break;
